+++ MidaiEsfe/MidaiEsfe.Aplicacion.EntidadesDeNegocio/TipoPersona.cs
@@ -8,7 +8,7 @@ namespace MidaiEsfe.Aplicacion.EntidadesDeNegocio
 {
     public class TipoPersona
     {
-        public byte Id { get; set; }
+        public Int64 Id { get; set; }
 
         public String Nombre { get; set; }
     }
